refactor: Completed header
