+++ src/components/Header/header.module.scss.d.ts
@@ -3,3 +3,4 @@ export const navigation: string;
 export const navigationLogo: string;
 export const navigationItem: string;
 export const panel: string;
+export const themeToggler: string;
