+++ src/styles/global.scss
@@ -1,6 +1,8 @@
 @import url('https://fonts.googleapis.com/css?family=Open+Sans:300,300i,400,400i,600,600i,700,700i,800,800i&display=swap');
 @import url('https://fonts.googleapis.com/icon?family=Material+Icons');
 
+@import './layout.scss';
+
 html {
   font-size: 10px;
 }
