+++ src/styles/layout.scss
@@ -80,17 +80,6 @@
   padding: 0.1rem;
 }
 
-.sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0,0,0,0);
-  border: 0;
-}
-
 ::placeholder {
   color: var(--black3);
 }
@@ -117,12 +106,6 @@ body:not(.dark-mode) .dark-mode-only {
   display: none;
 }
 
-.dark-mode-toggle {
-  border: none;
-  background: none;
-  cursor: pointer;
-  color: var(--color-text-accent);
-}
 
 /* Start: Syntax Highlight Overrides */
 
@@ -176,56 +159,6 @@ select:hover, select:active {
 
 [data-is-focused="true"] {
   outline: none;
-}
-
-.nav {
-  display: flex;
-  align-items: center;
-  flex-direction: row;
-  background: var(--color-fill-app);
-  border-bottom: var(--space-01) solid var(--black2);
-  height: var(--nav-height);
-  left: 0;
-  position: fixed;
-  right: 0;
-  top: 0;
-  z-index: 999;
-}
-
-.dark-mode .nav {
-  border: none;
-}
-
-.nav__logo {
-  height: var(--space-32);
-  margin-left: var(--space-24);
-}
-
-.nav__tabs__container {
-  margin-left: 120px;
-  list-style: none;
-  display: flex;
-  align-items: center;
-  max-width: 352px;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.nav .right-container {
-  display: flex;
-  padding-right: var(--space-24);
-  list-style: none;
-}
-
-.nav__tabs {
-  text-decoration: none;
-  margin-left: var(--space-16);
-}
-
-.nav__tabs > a {
-  text-decoration: none;
-  color: var(--color-text-primary);
-    padding: var(--space-24) var(--space-16)  calc(var(--space-24) - 2px) var(--space-16)
 }
 
 /* Start: Article Reader HTML Reset */
